# api/v1/REST/openapi.yaml
openapi: 3.0.3
info:
  title: Octi Sync Server
  description: |-
    This is the Octi Sync Server.
  contact:
    email: contact@jakob-moeller.de
  license:
    name: Apache 2.0
    url: https://github.com/jakobmoellerdev/octi-sync-server/blob/main/LICENSE
  version: v1
...
servers:
  - url: https://grylls.octi.darken.eu/v1
    description: The official Server!
  - url: http://localhost:8080/v1
    description: Run it Locally to try it out!
...
paths:
...
  /devices:
    get:
      tags:
        - devices
      summary: Get All registered Devices for your Account
      description: Receive Device Data for all devices in your Account
      operationId: getDevices
      parameters:
        - $ref: '#/components/parameters/XDeviceID'
      responses:
        '200':
          $ref: '#/components/responses/DeviceListResponse'
      security:
        - deviceAuth: []
...
components:
  parameters:
...
    XDeviceID:
      name: X-Device-ID
      in: header
      required: true
      description: |-
        Unique Identifier of the calling Device. If calling Data endpoints, must be presented in order
        to be properly authenticated.
      schema:
        $ref: '#/components/schemas/DeviceID'
...
  responses:
    DeviceListResponse:
      description: Device List containing multiple Devices
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DeviceList'
...
  schemas:
    DeviceID:
      type: string
      description: "Device ID is the unique identifier for a remote device"
      format: uuid
...
    ListItemCount:
      type: integer
      description: "Amount of Items contained in List"
    Device:
      type: object
      description: "a device"
      properties:
        id:
          $ref: '#/components/schemas/DeviceID'
      required:
        - id
    DeviceList:
      type: object
      description: "list of devices"
      properties:
        count:
          $ref: "#/components/schemas/ListItemCount"
        items:
          description: "array of devices, it will always at least contain the device of the authenticated user"
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Device'
      required:
        - count
        - items
...
  securitySchemes:
    deviceAuth:
      type: http
      scheme: basic